FIX: Fix tests not waiting for initializables with automaticallyAddRootGameObjectModules disabled
The module we inject using additive scenes never gets picked up in this
case. Instead, support resolving a Task when UnityKernel finishes
initializing and wait for them in the scene.

This also automatically supports multiple UnityKernels and
ContainerModuleLoaders in the scene, where the original approach was
racy because any one of them could resolve the callable the tests
inject first.