BUILD: Require Unity 6.1
This is technically a BC break but we already don't support or test on
Unity versions older than 6 and 6.1, so it's very likely already broken
in 2.0. The intent was to update but this was forgotten.

We could still target 6.0 but Test Framework 1.5 doesn't want to install
in 6.0 so let's just make it 6.1 and be done with it. We want to avoid
supporting multiple test framework versions since it's been in flux
the last few minor point releases and each version is wonky in its own
ways.